server/membership/migrations/0006_auto_20150811_0722.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import migrations
import phonenumber_field.modelfields


class Migration(migrations.Migration):
    dependencies = [
        ('membership', '0005_auto_20150811_0654'),
    ]

    operations = [
        migrations.AlterField(
            model_name='user',
            name='phone_number',
            field=phonenumber_field.modelfields.PhoneNumberField(max_length=20),
        ),
    ]
