requirements.txt
argparse==1.3.0
dj-database-url==0.3.0
django-extensions==1.5.6
django-filter==0.11.0
Django==1.8.4
djangorestframework==3.2.3
django-cors-headers==1.1.0
django-phonenumber-field==0.7.2
gunicorn==19.3
Markdown==2.6.2
newrelic==2.54.0.41
Pillow==2.9.0
psycopg2==2.6.1
pytz==2015.4
six==1.9.0
whitenoise==2.0.3
wsgiref==0.1.2
wheel==0.24.0
